requirements.txt
pyrogram
tgcrypto
Pillow
googletrans==3.1.0a0